% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_7day_forecast}
\alias{get_7day_forecast}
\title{Seven-Day descriptive forecast}
\usage{
get_7day_forecast(point_info)
}
\arguments{
\item{point_info}{object returned by \code{\link{get_geo_data}}}
}
\value{
a list of class "weather_forecast" containing the following components
\itemize{
  \item \code{@context} metadata
  \item \code{type} type of object
  \item \code{geometry} geometry, including coordinates
  \item \code{properties} information about the forecast
}
Importantly, the \code{properties} components is a list that whose
\code{$periods} element is a data frame containing the forecast.
}
\description{
Given an object with geographic information about a point (the output of
the function \code{get_geo_data}), return an object containing information
about the National Weather Service forecast for the next 7 days.
}
\examples{
ithaca_info <- get_geo_data( 42.440, -76.497 )
forecast <- get_7day_forecast( ithaca_info )
print(forecast)
}
