% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{get_geo_data}
\alias{get_geo_data}
\title{Retrieve geographic data for given longitude and latitude}
\usage{
get_geo_data(lat, lon)
}
\arguments{
\item{lat}{latitude in decimal degrees north (positive for northern hemisphere)}

\item{lon}{longitude in decimal degrees east (positive for eastern hemisphere)}
}
\value{
a list with the following components:
\itemize{
  \item \code{@context} metadata
  \item \code{id} original url accessed
  \item \code{type} type of object
  \item \code{geometry} geometry, including coordinates
  \item \code{properties} information needed to gather further weather
                          data related to this coordinate
}
}
\description{
Get the National Weather Service forecast grid information for a
given latitude and longitude coordinate.
}
\examples{
# get geographic information for Ithaca, NY
ithaca_info <- get_geo_data( 42.440, -76.497 )
print( ithaca_info$geometry )
print( ithaca_info$properties$relativeLocation$properties$city )
print( ithaca_info$properties$forecast )
}
