% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.weather_forecast}
\alias{print.weather_forecast}
\title{S3 method (function) to print the forecast}
\usage{
\method{print}{weather_forecast}(x, ..., nperiods = NULL)
}
\arguments{
\item{x}{a "weather_forecast" object, returned from
\code{get_7day_forecast}.}

\item{...}{further arguments passed to print}

\item{nperiods}{number of 1/2 day periods to print. Defaults to 2.}
}
\description{
Nicely print the information in a weather forecast object.
}
\examples{
ithaca_info <- get_geo_data( 42.440, -76.497 )
forecast <- get_7day_forecast( ithaca_info )
print(forecast, 2)
print(forecast, 4)
forecast
}
