% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstate_cities}
\alias{upstate_cities}
\title{Latitude/Longitude Coordinates of Cities in Upstate NY}
\format{
a dataframe with 8 rows and 3 columns
\describe{
    \item{name}{name of the city}
    \item{lat}{latitude of city, in decimal degrees north}
    \item{lon}{longitude of city, in decimal degrees east}
}
}
\usage{
upstate_cities
}
\description{
A dataset with the coordinates of 8 cities in Upstate NY
}
\keyword{datasets}
